% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{abridge_labels}
\alias{abridge_labels}
\title{Shorten labels to a glottocode substring}
\usage{
abridge_labels(phy)
}
\arguments{
\item{phy}{A phylo or multiPhylo object, containing one or more trees to
manipulate.}
}
\value{
A phylo or multiPhylo object, the manipulated tree(s).
}
\description{
Shortens tip and node labels to a glottocode substring within them. If any
labels lack a glottocode substring, a warning is given.
}
\details{
Glottocodes comprise four lowercase letters (or b10b or 3adt) followed by
four numbers, and are only identified if they are initial in the string or
are preceded by [.
}
