% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.R
\name{clone_node}
\alias{clone_node}
\title{Clone named nodes to self-daughter tips}
\usage{
clone_node(phy, label)
}
\arguments{
\item{phy}{A phlyo object}

\item{label}{A vector of strings,
the labels of the nodes to clone}
}
\value{
phy
}
\description{
A node is cloned once, even if listed
more than once in labels
}
