% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.R
\name{clone_tip}
\alias{clone_tip}
\title{Clone named tips}
\usage{
clone_tip(phy, label)
}
\arguments{
\item{phy}{A phylo object. The tree to manipulate.}

\item{labels}{A character vector containing tip labels.}
}
\value{
A phylo object containing the modified tree.

phy
}
\description{
#' Clones tips in a tree as self-sisters, under a new parent node.
}
\details{
To clone a tip, resulting in \eqn{n} copies, include its tip label \ean{n-1}
times in \code{label}.

Cloned tips will appear under a new node of the same name. Duplicate tips
will be assigned labels with a suffix \code{-1}, \code{-2}, \code{-3}, ...
}
