% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{edgeroot}
\alias{edgeroot}
\title{Edgeroot
The function edgeroot takes a rooted 
tree (input phy) and reroots that tree 
at the midpoint of the specified edge}
\usage{
edgeroot(x, phy)
}
\arguments{
\item{x}{An integer, specifying an egde.}

\item{phy}{A phylo object.}
}
\description{
Edgeroot
The function edgeroot takes a rooted 
tree (input phy) and reroots that tree 
at the midpoint of the specified edge
}
