% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_glottolog_families}
\alias{get_glottolog_families}
\title{A simple glottolog language family metadata table}
\usage{
get_glottolog_families(glottolog_version = get_newest_version())
}
\arguments{
\item{glottolog_version}{A character string. Which glottolog version to use.
Currently, available options are \code{'4.3'} and \code{'4.4'}.}
}
\description{
Returns a dataframe of glottolog metadata with columns: \code{tree},
\code{tree_name}, \code{n_tips}, \code{n_nodes} and \code{main_macroarea}.
}
\details{
By default, returns metadata from the newest version of glottolog. Obtain
metadata for older versions by setting the parameter
\code{glottolog_version}.
}
