% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_glottolog_languages}
\alias{get_glottolog_languages}
\title{A simple glottolog language metadata table}
\usage{
get_glottolog_languages(glottolog_version = get_newest_version())
}
\arguments{
\item{glottolog_version}{A character string. Which glottolog version to use.
Currently, available options are \code{'4.3'} and \code{'4.4'}.}
}
\description{
Returns a dataframe of glottolog metadata with columns: \code{glottocode},
\code{isocode}, \code{name}, \code{name_in_tree}, \code{position},
\code{tree} and \code{tree_name}.
}
\details{
By default, returns metadata from the newest version of glottolog. Obtain
metadata for older versions by setting the parameter \code{glottolog_version}.
}
