% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_glottolog_phylo_geo}
\alias{get_glottolog_phylo_geo}
\title{An extended glottolog metadata table}
\usage{
get_glottolog_phylo_geo(glottolog_version = get_newest_version())
}
\arguments{
\item{glottolog_version}{A character string. Which glottolog version to use.
Currently, available options are \code{'4.3'} and \code{'4.4'}.}
}
\description{
Returns a dataframe of glottolog geographical and phylogenetic metadata with
columns: \code{glottocode}, \code{name}, \code{vertex_name},
\code{family_name}, \code{isocodes}, \code{level}, \code{macroarea},
\code{latitude}, \code{longitude}, \code{vertex_type}, \code{vertex_label},
\code{tree} and \code{family_glottocode}.
}
\details{
By default, returns metadata from the newest version of glottolog. Obtain
metadata for older versions by setting the parameter
\code{glottolog_version}.
}
