% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.R
\name{get_glottolog_trees}
\alias{get_glottolog_trees}
\title{Glottolog trees by version}
\usage{
get_glottolog_trees(family, glottolog_version = get_newest_version())
}
\arguments{
\item{family}{A character vector. Elements are names of glottolog families
whose trees are to be returned. If \code{family} is left unspecified, all
trees are returned.}

\item{glottolog_version}{A character string. Which glottolog version to use.
Currently, available options are \code{'4.3'} and \code{'4.4'}.}
}
\value{
A multiPhylo object containing the glottolog trees.
}
\description{
Returns a multiPhylo object containing all, or a requested subset, of the
glottolog trees.
}
\details{
By default, trees are returned from the most recent version of glottolog.
Alternatively, an older version of glottolog can be specified.
}
