% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_newest_version}
\alias{get_newest_version}
\title{The current glottolog version}
\usage{
get_newest_version()
}
\value{
A character string.
}
\description{
Returns the newest version of glottolog for which data is included in this
package, which is \code{'4.4'}.
}
