% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{glottolog_languages}
\alias{glottolog_languages}
\title{Simple metadata table}
\usage{
glottolog_languages()
}
\value{
A dataframe with columns
  glottocode, isocode, name, 
  name_in_tree, position,
  tree, tree_name
}
\description{
Simple metadata table
}
