% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.R
\name{glottolog_supertree}
\alias{glottolog_supertree}
\title{Create a glottolog super-tree}
\usage{
glottolog_supertree(
  macro_groups = as.list(unique(glottolog_families()$main_macroarea))
)
}
\arguments{
\item{macro_groups}{A list}
}
\description{
Create a glottolog super-tree
}
