% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.R
\name{keep_tips}
\alias{keep_tips}
\title{Keep a set of nodes & tips, as tip set}
\usage{
keep_tips(phy, label)
}
\arguments{
\item{phy}{A phylo object}

\item{labels}{A vector of strings,
which are tip and node labels}
}
\description{
Keep a set of nodes & tips, as tip set
}
