% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{label_by_glottocode}
\alias{label_by_glottocode}
\title{Shorten tree labels to the glottocode}
\usage{
label_by_glottocode(phy)
}
\arguments{
\item{phy}{A phylo or multiPhylo object}
}
\value{
A phylo object
}
\description{
Retain only the glottocode substring
within the tip and node labels
}
