% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{label_by_name}
\alias{label_by_name}
\title{Change glottocode tip and node labels to names}
\usage{
label_by_name(phy)
}
\arguments{
\item{phy}{A phylo object}
}
\value{
A phylo object
}
\description{
Change glottocode tip and node labels to names
}
