% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{phylo_average}
\alias{phylo_average}
\title{Phylogenetically-weighted averages}
\usage{
phylo_average(phy = NULL, data = NULL)
}
\arguments{
\item{phy}{A phylo or multiPhylo object}

\item{data}{A dataframe}
}
\value{
A list, containing phy, data,
  and dataframes: ACL_weights, BM_weights,
  ACL_averages, BM_averages
}
\description{
Phylogenetically-weighted averages
}
