% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{relabel_with_names}
\alias{relabel_with_names}
\title{Change labels from glottocodes to names}
\usage{
relabel_with_names(phy, glottolog_version = get_newest_version())
}
\arguments{
\item{phy}{A phlyo object, the tree to manipulate.}

\item{glottolog_version}{A character string. Which glottolog version to use.
Currently, available options are \code{'4.3'} and \code{'4.4'}.}
}
\value{
A phlyo object, the manipulated tree.
}
\description{
Looks up glottolog language names corresponding to glottocodes and replaces
tip and node labels, which contain glottocodes, with the appropriate names.
}
\details{
Labels without glottocodes are left unchanged and a warning is given. The
version of glottolog to use for look-up can be controlled with
\code{glottolog_version}.
}
