% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.R
\name{select_tip}
\alias{select_tip}
\title{Select tips}
\usage{
select_tip(phy, label)
}
\arguments{
\item{phy}{A phylo object. The tree to manipulate.}

\item{label}{A character vector containing tip labels.}
}
\value{
A phylo object containing the modified tree.
}
\description{
From a tree, select which tips are to be kept.
}
\examples{

library(ape)

tree <- abridge_labels(get_glottolog_trees("Tangkic"))
plot(tree)
nodelabels(tree$node.label)

tree2 <- select_tip(tree, c("lard1243", "kang1283", "kaya1319"))
plot(tree2)
nodelabels(tree2$node.label)

## How to remove specific tips, keeping all others

all_tips <- tree$tip.label
all_tips
unwanted_tips <- "nyan1300"
selected_tips <- all_tips[!(all_tips \%in\% unwanted_tips)]
selected_tips
tree3 <- select_tip(tree, selected_tips)
plot(tree3)
nodelabels(tree3$node.label)

unwanted_tips <- c("nyan1300", "kang1283")
selected_tips <- all_tips[!(all_tips \%in\% unwanted_tips)]
selected_tips
tree4 <- select_tip(tree, selected_tips)
plot(tree4)
nodelabels(tree4$node.label)

# By convention, when all tips below a node are removed, the node is removed.
# Here we remove nyan1300 and kaya1319. Node kaya1318 is removed too:
unwanted_tips <- c("nyan1300", "kaya1319")
selected_tips <- all_tips[!(all_tips \%in\% unwanted_tips)] 
tree5 <- select_tip(tree, selected_tips)
plot(tree5)
nodelabels(tree5$node.label)

## How to remove sister tips, but keep the node above them as a tip

# step 1 - clone the node above as a new tip
tree5a <- clone_node(tree, "kaya1318")
plot(tree5a)
nodelabels(tree5a$node.label)

# step 2 - remove the sisters
all_tips <- tree5a$tip.label 
unwanted_tips <- c("nyan1300", "kaya1319")
tree5b <- select_tip(tree5a, all_tips[!(all_tips \%in\% unwanted_tips)])
plot(tree5b)
nodelabels(tree5b$node.label)

# step 3 - collapse the node above (but retain its clone, which is a tip)
tree5c <- collapse_node(tree5b, "kaya1318")
plot(tree5c)
nodelabels(tree5c$node.label)
}
