% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.R
\name{select_tips}
\alias{select_tips}
\title{Select tips from current tips and nodes}
\usage{
select_tips(phy, label)
}
\arguments{
\item{phy}{A phylo object. The tree to manipulate.}

\item{labels}{A character vector containing tip and node labels.}
}
\value{
A phylo object containing the modified tree.
}
\description{
From a tree, select existing tips to be kept, or to be cloned, and existing
internal nodes to be cloned as tips. To keep a tip, include its tip label
once in \code{label}. To clone a tip, resulting in \eqn{n} copies, includes
its tip label \ean{n} times in \code{label}. To clone an internal node as a
self-daughter tip, include its node label in \coded{label}.
}
\details{
Cloned tips will appear under a new node of the same name. Duplicate tips and
nodes will be assigned labels with a suffix \code{-1}, \code{-2}, \code{-3},
...
}
