% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_lengths.R
\name{set_branch_lengths_1}
\alias{set_branch_lengths_1}
\title{Set all branch length to 1}
\usage{
set_branch_lengths_1(phy)
}
\arguments{
\item{phy}{A phlyo object, the tree to manipulate.}
}
\value{
A phlyo object, the manipulated tree.
}
\description{
Sets all branch lengths in a tree to 1.
}
\examples{

library(ape)
tree <- abridge_labels(get_glottolog_trees("Tangkic"))
tree2 <- clone_tip(tree, "nyan1300", n = 2, subgroup = TRUE)
plot(tree2)
nodelabels(tree2$node.label)
tree3 <- set_branch_lengths_1(tree2)
plot(tree3)
nodelabels(tree3$node.label)
}
