% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_lengths.R
\name{set_branch_lengths_exp}
\alias{set_branch_lengths_exp}
\title{Exponentialise branch lengths}
\usage{
set_branch_lengths_exp(phy, ultrametric = FALSE)
}
\arguments{
\item{phy}{A phylo object}
}
\value{
A phylo object
}
\description{
Exponentialise branch lengths
}
