% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_lengths.R
\name{set_deepest_branch_lengths}
\alias{set_deepest_branch_lengths}
\title{Set length of deepest branches}
\usage{
set_deepest_branch_lengths(phy, branch_length = 1)
}
\arguments{
\item{phy}{A phlyo object, the tree to manipulate.}

\item{branch_length}{A numeric stating the branch length.}
}
\value{
A phlyo object, the manipulated tree.
}
\description{
Sets lengths of branches immediately below the root to the same,
user-specified length.
}
