% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_lengths.R
\name{set_first_branch_lengths}
\alias{set_first_branch_lengths}
\title{Set lengths of first branches}
\usage{
set_first_branch_lengths(phy, branch_length = 1)
}
\arguments{
\item{phy}{A phlyo object}
}
\value{
A phlyo object
}
\description{
Set lengths of first branches
}
