% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_lengths.R
\name{ultrametricise}
\alias{ultrametricise}
\title{Ultrametricise tree by stretching final edges}
\usage{
ultrametricise(phy)
}
\arguments{
\item{phy}{A phylo object}
}
\value{
A phylo object
}
\description{
Ultrametricise tree by stretching final edges
}
