% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{which_tree}
\alias{which_tree}
\title{Tree numbers of glottolog families}
\usage{
which_tree(family = NULL, glottolog_version = get_newest_version())
}
\arguments{
\item{family}{A character vector. Elements are names of glottolog families
whose trees are to be returned. If \code{family} is left unspecified, all
trees are returned.}

\item{glottolog_version}{A character string. Which glottolog version to use.
Currently, available options are \code{'4.3'} and \code{'4.4'}.}
}
\value{
A named vector of integers, giving the tree numbers and the
  family names as the vector names.
}
\description{
Returns the tree number of one or more glottolog families.
}
\details{
By default, returns metadata from the newest version of glottolog. Obtain
metadata for older versions by setting the parameter
\code{glottolog_version}.
}
