% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yin_2020_data}
\alias{yin_2020_data}
\title{Sonority sequencing violation data in 496 languages, 
from Yin (2020)}
\format{
A dataframe:
\describe{
  \item{name}{name of the lect}
  \item{glottocode}{glottocode of the lect}
  \item{onset_violation}{presence (1) or absence (0) of word-initial SSP violation}
  \item{coda_violation}{presence (1) or absence (0) of word-final SSP violation}}
}
\source{
Yin (2020)
}
\usage{
yin_2020_data
}
\description{
A dataframe containing language identifiers and data
on the presence of SSP violations in word-initial onsets
and word-final codas, as used in:
}
\details{
Yin, Ruihua. 2020. “Violations of the Sonority Sequencing 
Principle: How, and How Often?” 2020 Conference of the 
Australian Linguistic Society, Brisbane. 
https://als.asn.au/Conference/Program
}
\keyword{datasets}
