% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{ACL}
\alias{ACL}
\title{ACL weights}
\usage{
ACL(phy)
}
\arguments{
\item{phy}{A phylo object.}
}
\value{
A vector of weights.
}
\description{
Calculates phylogenetic weights (a.k.a. phylogenetic means) according to the classic method of Altschul, Carroll & Lipman DJ (1987).
}
