% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.R
\name{bind_as_rake}
\alias{bind_as_rake}
\title{Bind trees with high-level rake}
\usage{
bind_as_rake(phy)
}
\arguments{
\item{phy}{A multiphylo object}
}
\value{
A phylo object, a single tree
}
\description{
Bind trees with high-level rake
}
