% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{BM}
\alias{BM}
\title{BranchManager weights}
\usage{
BM(phy)
}
\arguments{
\item{phy}{A phylo object}
}
\value{
A vector of weights
}
\description{
Calculates phylogenetic weights according the to BranchManager method of Stone & Sidow (2007).
}
\details{

}
