% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.R
\name{clone_node}
\alias{clone_node}
\title{Clone named nodes to self-daughter tips}
\usage{
clone_node(phy, label)
}
\arguments{
\item{phy}{A phylo object. The tree to manipulate.}

\item{labels}{A character vector containing node labels.}
}
\value{
A phylo object containing the modified tree.
}
\description{
Clones internal nodes in a tree as self-daughter tips.
}
\details{
A node is cloned only once, even if its label appears more than once in
\code{label}.
}
