% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.R
\name{clone_node}
\alias{clone_node}
\title{Clone named nodes to self-daughter tips}
\usage{
clone_node(phy, label)
}
\arguments{
\item{phy}{A phylo object. The tree to manipulate.}

\item{label}{A character vector containing node labels.}
}
\value{
A phylo object containing the modified tree.
}
\description{
Clones internal nodes in a tree as self-daughter tips.
}
\details{
The length of any new branch, between node n and its new clone, is set equal
to the longest of the original branches directly below node n.
}
\examples{

library(ape)

tree <- 
  set_branch_lengths_exp(abridge_labels(get_glottolog_trees("Tangkic")))
plot(tree)
nodelabels(tree$node.label)
tree2 <- clone_node(tree, "sout2758")
plot(tree2)
nodelabels(tree2$node.label)

tree3 <- clone_node(tree, c("sout2758", "gang1267", "kaya1318"))
plot(tree3)
nodelabels(tree3$node.label)

\dontrun{
# Returns error if any element of `label` is not in `phy`
tree4 <-  clone_node(tree, c("sout2758", "xxxx1234"))
}
}
