% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.R
\name{clone_tip}
\alias{clone_tip}
\title{Clone named tips}
\usage{
clone_tip(phy, label)
}
\arguments{
\item{phy}{A phlyo object}

\item{label}{A vector of strings,
some of which might be tip labels
of phy}
}
\value{
phy
}
\description{
Clone named tips
}
