% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.R
\name{collapse_node}
\alias{collapse_node}
\title{Collapse a node rootwards}
\usage{
collapse_node(phy, label)
}
\arguments{
\item{phy}{A phylo object. The tree to manipulate.}

\item{labels}{A character vector containing node labels.}
}
\value{
A phylo object containing the modified tree.
}
\description{
Note this will also multichotomise all nodes with a 0-length branch above
them.
}
