% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.R
\name{collapse_node}
\alias{collapse_node}
\title{Collapse a node rootwards}
\usage{
collapse_node(phy, label)
}
\arguments{
\item{phy}{A phylo object}

\item{label}{A string, the labels
of the nodes to collapse}
}
\value{
A phylo object
}
\description{
Note this will also multichotomise
all nodes with a 0-length branch above 
them
}
