% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{extract_glottocode}
\alias{extract_glottocode}
\title{Extract parts from a glottolog tree label}
\usage{
extract_glottocode(labels)
}
\arguments{
\item{label}{A string}
}
\value{
A string
}
\description{
Extract parts from a glottolog tree label
}
