% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_lengths.R
\name{set_branch_lengths_1}
\alias{set_branch_lengths_1}
\title{Set all branch length to 1}
\usage{
set_branch_lengths_1(phy)
}
\arguments{
\item{phy}{A phlyo object, the tree to manipulate.}
}
\value{
A phlyo object, the manipulated tree.
}
\description{
Sets all branch lengths in a tree to 1.
}
