% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_lengths.R
\name{set_branch_lengths_1}
\alias{set_branch_lengths_1}
\title{Set branch length to 1}
\usage{
set_branch_lengths_1(phy)
}
\arguments{
\item{phy}{A phylo object}

\item{ultrametric}{A logical. Whether or not
to lengthen branches immediately above
tips to make the tree ultrametric}
}
\value{
A phylo object
}
\description{
Set branch length to 1
}
