% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_lengths.R
\name{set_branch_lengths_exp}
\alias{set_branch_lengths_exp}
\title{Exponentialize branch lengths}
\usage{
set_branch_lengths_exp(phy)
}
\arguments{
\item{phy}{A phlyo object, the tree to manipulate.}
}
\value{
A phlyo object, the manipulated tree.
}
\description{
Sets the deepest branches to length 1/2, the next deepest to 1/4, the next to
1/8, etc.
}
\examples{

library(ape)
tree <- abridge_labels(get_glottolog_trees("Siouan"))
plot(tree)
nodelabels(tree$node.label)
tree2 <- set_branch_lengths_exp(tree)
plot(tree2)
nodelabels(tree2$node.label)
}
